backend/mesh-generator/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
archspec==0.2.5
async-timeout==5.0.1
attrs==25.3.0
boltons==23.1.0
Brotli==1.0.9
brotlipy==0.7.0
certifi
cffi
charset-normalizer==3.4.6
chumpy==0.70
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
cycler==0.12.1
distro==1.8.0
filelock==3.16.1
flatbuffers==25.12.19
flatten-dict==0.4.2
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
fvcore==0.1.5.post20221221
hf-xet==1.4.2
hpack==4.0.0
huggingface_hub==0.36.2
humanfriendly==10.0
idna==3.4
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.10
Jinja2==3.1.6
joblib==1.4.2
jsonpatch==1.32
jsonpointer==2.3
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
kornia==0.7.3
lazy_loader==0.4
lightning-utilities==0.11.9
llvmlite==0.41.1
loguru==0.7.3
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
multidict==6.1.0
networkx==3.1
numba==0.58.1
numpy==1.23.5
onnxruntime==1.19.2
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
packaging==26.0
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
portalocker==3.0.0
propcache==0.2.0
protobuf==5.29.6
pybind11==3.0.2
pycparser==2.21
PyMatting==1.1.14
pyparsing==3.1.4
python-dateutil==2.9.0.post0
PyWavelets==1.4.1
PyYAML==6.0.3
rembg==2.0.61
requests==2.32.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sympy==1.13.3
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==1.0.25
tinyobjloader @ git+https://github.com/syoyo/tinyobjloader.git@966edceaf8cdca7996c4e9a1c5ced2938de63366
torch==1.13.1
torchaudio==0.13.1
torchmetrics==1.5.2
torchvision==0.14.1
tqdm==4.67.3
trimesh==4.11.4
typing_extensions==4.13.2
urllib3==2.2.3
yacs==0.1.8
yarl==1.15.2